Add Hydration Reminder sample app
The Hydration Reminder app make sure you are getting enough water
and afforded the common health consequences of dehydration,
like headache, dizziness or not be able to shed a tear when yout test are passing.
It works by sending you a reminder notification to drink water every 15 minutes when your phone is charging.